switch between list and buddy working, free list not working

+++ Kernel/include/memoryManager.h
@@ -3,9 +3,10 @@
 
 #include <stddef.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 void initializeMemoryManager(char * heap_base, size_t heap_size);
-void *malloc(size_t nbytes);
+void *malloc(uint64_t nbytes);
 void free(void *block);
 void memoryDump();
